refactor: split Text atom style, type, component
